Update testing for Cpp

+++ project1/Cpp/src/main.cpp
@@ -9,10 +9,10 @@
 using json = nlohmann::json;
 
 void printProgress(double percentage) {
-    int val = (int) (percentage * 100);
+    double val = percentage * 100;
     int lpad = (int) (percentage * PBWIDTH);
     int rpad = PBWIDTH - lpad;
-    printf("\r%3d%% [%.*s%*s]", val, lpad, PBSTR, rpad, "");
+    printf("\r%4.1f%% [%.*s%*s]", val, lpad, PBSTR, rpad, "");
     fflush(stdout);
 }
